Add support for port 8080
This adds support for accessing the containers via port 8080 for use with the
flask application. This required and an adjustment where we just use the last
two digits of the port to signify which port the container uses. The ports
prefixed with 330XX are for ssh and ports prefixed with 340XX are for the flask
application port.

In addition, I'm removing the cpu and memory constraints as they seem to be what
was causing the problem before at the last DOBC.