Setup MySQL containers for each shell container
This creates a MariaDB (MySQL) container for each shell container and links it
up using docker. MySQL has been tuned so that it uses the least amount of memory
as possible. I also added a script to start individual MySQL containers as
needed.

The dobc-centos container has an alias for the 'mysql' command which
automatically uses the environment variables which the MariaDB container has for
connecting up without needing to know the password. In addition, I've bumped up
the flavor to match the size that we use in production to ensure we have enough
memory for 100 instances of each.